APIs/funtranslations.com/index/2.3/python-aiohttp-client/openapi_server/controllers/starwars_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def translate_cheunh_get(request: web.Request, text) -> web.Response:
    """translate_cheunh_get

    Translate from English to Starwars cheunh.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)


async def translate_gungan_get(request: web.Request, text) -> web.Response:
    """translate_gungan_get

    Translate from English to Starwars Gungan Language.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)


async def translate_huttese_get(request: web.Request, text) -> web.Response:
    """translate_huttese_get

    Translate from English to Starwars Huttese Language.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)


async def translate_mandalorian_get(request: web.Request, text) -> web.Response:
    """translate_mandalorian_get

    Translate from English to Starwars Mandalorian Language.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)


async def translate_sith_get(request: web.Request, text) -> web.Response:
    """translate_sith_get

    Translate from English to Sith Speak.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)


async def translate_yoda_get(request: web.Request, text) -> web.Response:
    """translate_yoda_get

    Translate from English to Yoda Speak.

    :param text: Text to translate
    :type text: str

    """
    return web.Response(status=200)
